cli_main update
